+++ game/components/c_camera.h
@@ -9,8 +9,16 @@
 namespace Verse::Component
 {
     struct Camera {
-        Vec2f centre;
+        Vec2f pos;
         Vec2f vel;
+        
+        Rect2 bounds;
+        Vec2 focus_size;
         float l,r,t,b;
+        
+        float look_ahead;
+        float la_speed;
+        
+        std::function<bool()> controller;
     };
 }