+++ game/systems/s_camera.h
@@ -13,10 +13,8 @@
 
 namespace Verse::System::Camera
 {
-    void init(Component::Camera* camera, Vec2f pos, Vec2f size);
+    void init(Component::Camera* camera);
     void setActive(Component::Camera* camera);
-    void setActive(Component::Camera* camera, Rect2 cam_bounds);
-    void setActive(Component::Camera* camera, Scene &scene);
     void update(Config &c, Scene& scene);
     void move(Config &c, Vec2f pos, int input);
     void updatePoints(Config &c, Vec2f &pos);