+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 
 ## **0.2** ~ Light
 
-### **[0.1.3]** ~ Serialization _2021-04-06_
+### **[0.1.3]** ~ Serialization _2021-04-07_
 #### Added
 - YAML Serialization
 - Load, write, destroy YAML files
 - Edit/remove particular entries
 - Append with specific typings
 - Scene serialization
+- Player serialization
 #### Changed
 - Different Vec2 and Vec2f, Rect2 and Rect2f (For precission and cleanness matters)
 - Rect2 now uses x, y, w, h instead of pos and size